UBUNTU: [Config] Enable virtio-net as built-in to avoid race
Currently cloud-init is racy, and may or may not come up correctly
when virtio-net is a module, and the datasource is accessible via
virtio-net. To ensure minimal images are robust on Canonical
Distribution of Openstack, mark virtio-net as built-in. We can try to
modularize virtio-net again, if and when cloud-init race is resolved.

BugLink: https://bugs.launchpad.net/bugs/2036968

Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@canonical.com>
[ arighi: add LP reference to annotations ]
Signed-off-by: Andrea Righi <andrea.righi@canonical.com>