UBUNTU: [Packaging] remove debian.*/variants
Now 'variants' is no longer used for building packages.

It is true that the 'variants' is used in linux-meta, but there is
no good reason to have it in linux-main.

Only the remaining function is to check the variants from
'debian/rules printenv', but it is not so useful.

Remove it.

Signed-off-by: Masahiro Yamada <masahiro.yamada@canonical.com>
Acked-by: Dimitri John Ledkov <dimitri.ledkov@canonical.com>
Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@canonical.com>