UBUNTU: [Config] move qcom clk and serial options as builtin
BugLink: https://bugs.launchpad.net/bugs/2083559

Enable Qualcomm CLK, SC_GCC_7280 and SERIAL_QCOM_GENI options as builtin
for functional UART on Qualcomm RB3Gen2 platforms on normal boot and
with earlycon.

As modules it is impossible to debug early boot issues, as it relies on
a functional kernel and boot process.

Signed-off-by: Ricardo Salveti <ricardo.salveti@oss.qualcomm.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Guoqing Jiang <guoqing.jiang@canonical.com>
Signed-off-by: Roxana Nicolescu <roxana.nicolescu@canonical.com>