UBUNTU: [Packaging] perf: enable debuginfod support
For symbols resolution, the perf tool can use debuginfod server as source
of debuginfo files. This support can be requested through the argument
`--debuginfod` of `perf record` command. For example, to use Ubuntu debuginfod
server:

perf record --debuginfod='https://debuginfod.ubuntu.com' -p <pid>

This support requires the perf tool to be built with libdebuginfod
library. However, this library is not a build dependency of the kernel
currently.
This patch adds libdebuginfod-dev as build dependency of the kernel
so that the debuginfod support is enabled for `perf`.

Signed-off-by: Hector Cao <hector.cao@canonical.com>
Signed-off-by: Timo Aaltonen <timo.aaltonen@canonical.com>