UBUNTU: [Packaging] linux-lib-rust: Fix incorrect package name
This package should be in the same namespace as the other libraries and
headers packages, i.e., it should always be named

linux-lib-rust-<version>-<flavor>

and *not*

linux-unstable-lib-rust-<version>-<flavor> or
linux-aws-lib-rust-<version>-<flavor> or
...

This change only affects derivatives and unstable. The package name
produced by the main linux source remains the same.

While at it, remove the unused variable 'indep_lib_rust_pkg_name'.

Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>
Acked-by: Agathe Porte <agathe.porte@canonical.com>
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>