UBUNTU: [Packaging] linux-bpf-dev: Skip packaging for additional kernel flavors
The linux-bpf-dev package is a broken abomination :-( It provides
vmlinux.h which is generated from vmlinux but some kernels have multiple
flavors so which one to use?? It should really be a flavored package...

Anyways, for now just ensure that we do all the packaging install and
binary steps only when processing the main/first flavor and skip them for
any additional flavors. Otherwise we end up with a packaged vmlinux.h that
is generated from the last flavor vmlinux which is likely not what we want.

While at it, add the missing dh_prep step.

Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>