UBUNTU: [Packaging] bpftool: move to separate unversioned and unflavored binary package
BugLink: https://bugs.launchpad.net/bugs/2007308

The bpftool binary is used as build time dependency for multiple
Debian packages.

Before this patch, bpftool was provided by the linux-tools-common
package, which in reality shipped a shell script that attempted to
locate the real binary in /usr/lib/linux-tools/$(uname -r)/,
falling back to printing a human friendly message with
instructions for what package to install.

This does not work well for multiple reasons:
* Automated build pipelines are more often than not run in
  containers.
* Said containers usually host a user space filesystem that do
  not contain nor have any way to install linux-tools-$(uname -r)
  package.
* Automated build pipelines declaring bpftool as a build time
  dependency expect the functionality of bpftool when executing
  bpftool, and have no way of interpreting the human friendly
  message.
* This deviation from Debian creates perpetual per release manual
  work for Ubuntu for any Debian package depending on bpftool.

The BPF Compile Once - Run Everywhere (BPF CO-RE) concept revolves
around using bpftool to dump BPF Type Format (BTF) information from
the running kernel.

With this in mind, it comes across as counter intuitive that the
bpftool binary needs to be distributed in a versioned and
flavoured package.

Move bpftool into a separate unversioned and unflavoured binary
package.

Prefix bpftool package version with the bpftool projects own
versioning scheme.  This is aligned with the Debian packaging [0]
and expectations created by upstream stand-alone distribution of
bpftool [1].

Note that there are mulitple styles of variable use and gating in
the package source, and every change in this patch is made matching
local style.

0: https://salsa.debian.org/kernel-team/linux/-/commit/b20f5b55eb30c2ba0e1fa8f5d241370225063321
1: https://lore.kernel.org/bpf/267a35a6-a045-c025-c2d9-78afbf6fc325@isovalent.com/
Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>
Signed-off-by: Timo Aaltonen <timo.aaltonen@canonical.com>